task: update unit test
