+++ packages/java/neo-auth-service/src/test/java/org/neocoder/services/auth/AuthServiceTest.java
@@ -1,10 +1,11 @@
+package org.neocoder.services.auth;
+
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
-import org.neocoder.services.auth.AuthService;
 
 class AuthServiceTest {
